requirements.txt
Django==5.1.3
djangorestframework==3.15
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn
openai==1.53.0
httpx==0.27.2
drf-yasg==1.21.7
setuptools==80.0.0